+++ libs/bitfield-c/src/canutil/write.c
@@ -9,12 +9,12 @@ uint64_t float_to_fixed_point(const float value, const float factor,
         // round up to avoid losing precision when we cast to an int
         // TODO do we need a way to encode an int back to a signal without any
         // rounding?
-        raw += 0.5;
+        raw += (float)0.5;
     }
     return (uint64_t)raw;
 }
 
-uint64_t eightbyte_encode_float(float value, unsigned int bit_offset, unsigned int bit_size,
+uint64_t eightbyte_encode_float(float value, uint8_t bit_offset, uint8_t bit_size,
         float factor, float offset) {
     uint64_t result = 0;
     if(!eightbyte_set_bitfield(float_to_fixed_point(value, factor, offset),
@@ -24,14 +24,14 @@ uint64_t eightbyte_encode_float(float value, unsigned int bit_offset, unsigned i
     return result;
 }
 
-uint64_t eightbyte_encode_bool(const bool value, const unsigned int bit_offset,
-        const unsigned int bit_size) {
+uint64_t eightbyte_encode_bool(const bool value, const uint8_t bit_offset,
+        const uint8_t bit_size) {
     return eightbyte_encode_float(value, bit_offset, bit_size, 1.0, 0);
 }
 
-bool bitfield_encode_float(const float value, const unsigned int bit_offset,
-        const unsigned int bit_size, const float factor, const float offset,
-        uint8_t destination[], const unsigned int destination_length) {
+bool bitfield_encode_float(const float value, const uint8_t bit_offset,
+        const uint8_t bit_size, const float factor, const float offset,
+        uint8_t destination[], const uint32_t destination_length) {
     if(!set_bitfield(float_to_fixed_point(value, factor, offset), bit_offset,
                 bit_size, destination, destination_length)) {
         // debug("%f will not fit in a %d bit field", value, bit_size);
@@ -40,9 +40,9 @@ bool bitfield_encode_float(const float value, const unsigned int bit_offset,
     return true;
 }
 
-bool bitfield_encode_bool(const bool value, const unsigned int bit_offset,
-        const unsigned int bit_size, uint8_t destination[],
-        const unsigned int destination_length) {
+bool bitfield_encode_bool(const bool value, const uint8_t bit_offset,
+        const uint8_t bit_size, uint8_t destination[],
+        const uint32_t destination_length) {
     return bitfield_encode_float(value, bit_offset, bit_size, 1.0, 0,
             destination, destination_length);
 }
